% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestOfTwo.R
\name{BestOfTwo}
\alias{BestOfTwo}
\title{BestOfTwo}
\usage{
BestOfTwo(Clust1, Clust2, InstanceList)
}
\description{
Looks at the resulting clustering from the two algorithms and picks the best
}
\details{
@param Clust1 Clustering from PickBestCluster
 @param Clust2 Clustering from CC_Pivot 
 @param InstanceList List of all clusterings
}

