% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestOneElementMove.R
\name{BestOneElementMove}
\alias{BestOneElementMove}
\title{BestOneElementMove}
\usage{
BestOneElementMove(Wp, InstanceList, C)
}
\arguments{
\item{Wp}{positive consensus matrix from CC-Pivots initialization}

\item{InstanceList}{List of clustering instances}

\item{C}{Initial clustering}
}
\description{
Performs Best one element move algorithm
}

