% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_Pivot.R
\name{CC.Pivot}
\alias{CC.Pivot}
\title{CC_Pivot}
\usage{
CC.Pivot(G)
}
\arguments{
\item{G}{A graph object (list) with elements V, Ep and Em}
}
\description{
This function performs consensus clustering with CC_Pivot algorithm
}

