% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster2Adj.R
\name{Cluster2Adj}
\alias{Cluster2Adj}
\title{Cluster2Adj}
\usage{
Cluster2Adj(ClusterList)
}
\arguments{
\item{ClusterList}{All clusterings from one clustering instance}
}
\description{
Converts a cluster to an adjacency matrix
}

