% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedClusteringWithReps.R
\name{CombinedClusteringWithReps}
\alias{CombinedClusteringWithReps}
\title{CombinedClusteringWithReps}
\usage{
CombinedClusteringWithReps(InstanceList, Reps)
}
\arguments{
\item{InstanceList}{List of clusters}

\item{Reps}{Number of repetations to do for CC-Pivot}
}
\description{
Performs best of two clustering with reps and picks the best clustering
}

