% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conver2Labels.R
\name{Convert2Labels}
\alias{Convert2Labels}
\title{Convert2Labels}
\usage{
Convert2Labels(Instance, N)
}
\arguments{
\item{Instance}{Clustering instance returned by CC-pivot}

\item{N}{Total number of elements}
}
\description{
Converting list returned by CC-pivot to Labels
}

