% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr2Cons.R
\name{Corr2Cons}
\alias{Corr2Cons}
\title{Corr2Cons function}
\usage{
Corr2Cons(InstanceList)
}
\arguments{
\item{InstanceList}{List of all clusterings}
}
\description{
This function all clustering instances and converts to consensus clustering matrices
}

