% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClusterDistance.R
\name{GetClusterDistance}
\alias{GetClusterDistance}
\title{Get.Cluster.Distance}
\usage{
GetClusterDistance(Clust1, Clust2)
}
\arguments{
\item{Clust1}{First clustering}

\item{Clust2}{Second clustering}
}
\description{
This function calculates the distances between two clusterings
}

