% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PickBestCluster.R
\name{PickBestCluster}
\alias{PickBestCluster}
\title{PickBestCluster}
\usage{
PickBestCluster(InstanceList)
}
\arguments{
\item{InstanceList}{List of all clusterings}
}
\description{
This function picks the clustering with lowest total distance
}

